Modificación de tipo dato recibido del endpoint y se realiza su conversión a entero y además se agrega método que recibe los productos en MainActivity

+++ dataAccess/src/main/java/com/ceiba/dataaccess/anticorruption/ProductTranslator.kt
@@ -13,7 +13,7 @@ object ProductTranslator {
         ProductBuilder().apply {
 
         setTitle(productResponse.title)
-        setPrice(productResponse.price)
+        setPrice(productResponse.price.toInt())
         setImageProduct(productResponse.imageProduct)
         setAddress(fromAddressApiToAddress(productResponse.address))
         setSpecs(fromSpecsApiToSpecs(productResponse.specs))
